Add autorization check. Middlewares. Correct appearance.

+++ resources/views/admin/categories/index.blade.php
@@ -6,14 +6,8 @@
     <!-- Content Header (Page header) -->
     <section class="content-header">
       <h1>
-        Blank page
-        <small>it all starts here</small>
+        Категории
       </h1>
-      <ol class="breadcrumb">
-        <li><a href="#"><i class="fa fa-dashboard"></i> Home</a></li>
-        <li><a href="#">Examples</a></li>
-        <li class="active">Blank page</li>
-      </ol>
     </section>
 
     <!-- Main content -->
@@ -21,9 +15,6 @@
 
       <!-- Default box -->
       <div class="box">
-            <div class="box-header">
-              <h3 class="box-title">Листинг сущности</h3>
-            </div>
             <!-- /.box-header -->
             <div class="box-body">
               <div class="form-group">
@@ -42,12 +33,15 @@
                   <tr>
                     <td>{{$category->id}}</td>
                     <td>{{$category->title}}</td>
-                    <td><a href="{{route('categories.edit', $category->id)}}" class="fa fa-pencil"></a> 
-                    {{Form::open(['route' => ['categories.destroy', $category->id], 'method' => 'delete'])}}
-                      <button type="submit" class="delete" onclick="return confirm('Вы уверены?');">
-                        <a class="fa fa-remove"></a>
-                      </button>
-                    {{Form::close()}}
+                    <td>
+                    @if(\Auth::user()->is_admin == 1)
+                      <a href="{{route('categories.edit', $category->id)}}" class="fa fa-pencil"></a> 
+                      {{Form::open(['route' => ['categories.destroy', $category->id], 'method' => 'delete'])}}
+                        <button type="submit" class="delete" onclick="return confirm('Вы уверены?');">
+                          <a class="fa fa-remove"></a>
+                        </button>
+                      {{Form::close()}}
+                    @endif
                     </td>
                   </tr>
                 @endforeach
